refactor: 이동 전략 각 기물로 이동
