--- src/main/java/janggi/domain/movestrategy/RelativePosition.java
@@ -1,4 +1,0 @@
-package janggi.domain.movestrategy;
-
-public record RelativePosition(int dx, int dy) {
-}
