+++ src/main/java/janggi/domain/piece/CannonPiece.java
@@ -1,12 +1,33 @@
 package janggi.domain.piece;
 
-import janggi.domain.movestrategy.CannonStrategy;
 import janggi.domain.board.Position;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 public class CannonPiece extends Piece {
     public CannonPiece(Team team) {
-        super(team, Name.CANNON, new CannonStrategy());
+        super(team, Name.CANNON);
+    }
+
+    @Override
+    public boolean canMoveByBasicMovingRule(Position from, Position to) {
+        return (from.isSameX(to) && !from.isSameY(to))
+                || (!from.isSameX(to) && from.isSameY(to));
+    }
+
+    @Override
+    public List<Position> findPath(Position from, Position to) {
+        List<Position> path = new ArrayList<>();
+        int stepX = Integer.compare(to.x(), from.x());
+        int stepY = Integer.compare(to.y(), from.y());
+        Position currentPosition = from;
+
+        while (!currentPosition.equals(to)) {
+            currentPosition = currentPosition.moveBy(stepX, stepY);
+            path.add(currentPosition);
+        }
+        return path;
     }
 
     @Override
