+++ src/main/java/janggi/domain/piece/SoldierPiece.java
@@ -1,14 +1,25 @@
 package janggi.domain.piece;
 
 import janggi.domain.board.Position;
-import janggi.domain.movestrategy.ChoSoldierStrategy;
-import janggi.domain.movestrategy.HanSoldierStrategy;
-import janggi.domain.movestrategy.MoveStrategy;
+import java.util.List;
 import java.util.Map;
 
 public class SoldierPiece extends Piece {
     public SoldierPiece(Team team) {
-        super(team, Name.SOLDIER, createMoveStrategy(team));
+        super(team, Name.SOLDIER);
+    }
+
+    @Override
+    public boolean canMoveByBasicMovingRule(Position from, Position to) {
+        if (getTeam() == Team.HAN) {
+            return canMoveHanSoldier(from, to);
+        }
+        return canMoveChoSoldier(from, to);
+    }
+
+    @Override
+    public List<Position> findPath(Position from, Position to) {
+        return List.of(to);
     }
 
     @Override
@@ -22,10 +33,17 @@ public class SoldierPiece extends Piece {
         return true;
     }
 
-    private static MoveStrategy createMoveStrategy(Team team) {
-        if (team == Team.HAN) {
-            return new HanSoldierStrategy();
+    private boolean canMoveHanSoldier(Position from, Position to) {
+        if (from.deltaY(to) == 1 && from.isSameX(to)) {
+            return true;
         }
-        return new ChoSoldierStrategy();
+        return from.distanceX(to) == 1 && from.isSameY(to);
+    }
+
+    private boolean canMoveChoSoldier(Position from, Position to) {
+        if (from.deltaY(to) == -1 && from.isSameX(to)) {
+            return true;
+        }
+        return from.distanceX(to) == 1 && from.isSameY(to);
     }
 }
