+++ src/test/java/janggi/domain/piece/GuardPieceTest.java
@@ -3,9 +3,6 @@ package janggi.domain.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.domain.board.Position;
-import janggi.domain.movestrategy.ElephantStrategy;
-import janggi.domain.movestrategy.GeneralStrategy;
-import janggi.domain.movestrategy.GuardStrategy;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -26,7 +23,8 @@ class GuardPieceTest {
     })
     void testMoveGeneral(int preX, int preY, int nextX, int nextY) {
         GeneralPiece generalPiece = new GeneralPiece(Team.HAN);
-        assertThat(generalPiece.canMoveByBasicMovingRule(new Position(preX, preY), new Position(nextX, nextY))).isTrue();
+        assertThat(
+                generalPiece.canMoveByBasicMovingRule(new Position(preX, preY), new Position(nextX, nextY))).isTrue();
     }
 
     @ParameterizedTest
@@ -39,7 +37,8 @@ class GuardPieceTest {
     })
     void testNotMovableGeneral(int preX, int preY, int nextX, int nextY) {
         GeneralPiece generalPiece = new GeneralPiece(Team.HAN);
-        assertThat(generalPiece.canMoveByBasicMovingRule(new Position(preX, preY), new Position(nextX, nextY))).isFalse();
+        assertThat(
+                generalPiece.canMoveByBasicMovingRule(new Position(preX, preY), new Position(nextX, nextY))).isFalse();
     }
 
     @ParameterizedTest
